Remove functions superceded by consult
consult-completing-yank is not visually as nice as my function, but
I'll endeavor to improve the one in consult instead of using mine.

Also, I need to make consult-line and consult-yank-* start in list
mode in embark occur. That can be done with advice, maybe that's the
way to go for now.