Indent cl-labels and cl-flet using common-lisp-indent-function
Thanks to @clemera for showing me @kriyative's hybrid-lisp-indent-function which has the great idea to use common-lisp-indent-function for some constructs and lisp-indent-function for others. This is a slightly simplified version of that function. Besides a little simplification I also modified it to only recognize the cl-lib functions and not also the cl ones.