Remove higlighting of first line with colon in markdown-mode
I mostly use markdown for reddit & GitHub so my markdown buffers almost always start directly with text, not metadata. Markdown-mode interprets a first line with a colon as metadata (sometimes) and highlights it. This change removes the higlighting.

The "proper" solution is instead to tighten up its definition of metadata so that it only recongizes keys I'd actually use, but since I don't actually write markdown with metadata, that would be overkill.