Replace consult-buffer project source with project-files
The project file source that comes with consult is the subset of recent files which are under the current project root. I think I prefer the source to list all files in the current project, as return by project-files. I'll keep an eye out for performance and maybe roll this back.