Let's try @karthink's this-buffer target idea
I put the target last so it doesn't get in the way but I can quickly
access it with a negative prefix.

Note that with the advice to rename-file and delete-file in the
previous commit, those actions here are more useful than they seem. If
the advice turns out to be too dangerous and I get used to these
actions I'll write some {delete,rename}-file-and-buffer commands.