Add back binding for find-library
I had removed all but one key binding for library commands, leaving only
C-c l for load-library. The rest of the commands I was using as embark actions from load-library, but I think find-library is common enough for me, that I should have a direct binding for it as well, as a second entry point to the library actions.

I don't often use consult-focus-lines except in embark collect buffers, and there I have it bound to F, so I'll change the global binding to C-c F and leave C-c f for find-library.