Use @karthink's version of the embark-act action for avy
I had not noticed that my version wasn't correctly restoring the
window when acting on a target in a different window.

I don't understand the purpose of returning t, though. And a quick
look at the avy source did not illuminate me, nor did I see any
difference when I removed it. I'll keep it and figure out what it's
for at some point (several, but not all of the built-in avy dispatch
actions return t like that).