Switch from SLY to SLIME to use my standard completion UI
SLY defaults to using its own funky (and slightly buggy) completion
UI, and I would rather use either consult-completion-in-region or
corfu-mode. I was not able to get SLY to use either of those for
completion, despite using SLY's capf functions. SLIME is old fashioned
and doesn't directly use completion-in-region-function, but one can
override the function slime-display-or-scroll-completions to do so.

I don't think I used any functionality in SLY that is nto present in
SLIME, except maybe the #v notation for previous results, but I almost
always used that just for the last result anyway...