Better directory guessing for ffap in eshell buffers
In eshell buffers, by default ffap only looks for files in the current
default-directory. This little library has it also look for the
directory heading in a multi-directory ls listing, and for a directory
at the end of the command that listed the file.

These heuristics are not perfect, but they find the right file in more
cases ---and in the process make Embark file actions much more useful
in eshell buffers.