Have symbols in Slime repl buffers classified as identifiers
This might be better to fix in Embark. The current rule is that the
symbol at point is considered to be an Emacs Lisp symbol if (1) the
buffer is in emacs-lisp-mode or (2) the buffer is in a mode that does
not derive from prog-mode and the symbol has been interned in Emacs.

The second clause is intended to catch Emacs Lisp symbols in help
buffers, info, and text where you are talking about Emacs: emails,
markdown buffers to paste into reddit (:P), your org mode notes,
a telega buffer, etc. But it accidentally also covers the case of an
indentifier in a repl than happens to be named like an interned Emacs
symbol.

I'll experiment with tighter rules for (2). And, also, broaden (1) to
cover ielm! I was shocked to learn the other day that some people like
it. :P