Go back to consult-completion-in-region
Corfu is very pretty and I do like that it pops up right at point, but
in the end I missed acting on completions more than I enjoyed the UI.

Corfu itself does cover the actions I want maybe 80% of the time: it
has corfu-show-location and corfu-show-documentation, which are great.
But I also occasionally want to use embark-collect-snapshot, or to put
several completion candidates on the kill-ring, or to mass insert
them (embark-act-all, embark-insert), or to look them up in the info
manual instead of describe-{function,variable}, or, as a special case
for email addresses, I want to use my remove-email-from-ecomplete
function to delete the email from my addressbook.

Maybe I should try the corfu-move-to-minibuffer command from the corfu
wiki, but honestly it just feels like an unnecessary extra step that
would be only justified if I disliked the consult-completion-in-region
UI and wanted to avoid it, but I was already perfectly used to it
before and I think I don't mind going back (we'll see). The key thing
is the consult preview: thanks to that, I often don't need to glance
down at the minibuffer!