Add command to pop-up buffer for text to send to clipboard
I'll use this for text I need to input in the browser (reddit, GitHub,
even WhatsApp and Telegram). The buffer that pops up is in markdown
mode and has the TeX input method enabled. A simple C-c C-c kills the
buffer and window and copies the text to the clipboard.