New elfeed-comments command
The RSS feeds for Hacker News and Reddit have a useful link named
"comments". I wrote the embark-elfeed-target-first-url-in-content
target finder for that link, but (1) it isn't the first link in reddit
feeds, (2) it's overly general: I don't want the first link in other
cases, (3) I only ever used the "browse in eww" action so a dedicated
command is more ergonomic.