New vc-git-commit command (mainly for merges in case of conflict)
When there are conflicts, VC lets you know so you can fix them, and
when you do it automatically stages the files, but you can't actually
finish the merge with vc-next-action. I think it's because that calls
git with --only list-of-files, and for concluding a merge git wants
you to call it with no file names. So far, I've been using magit after
resolving the conflicts to finish the merge, but this silly little
vc-git-commit command is all I really need.