Uniformize Vertico and MCT directory deleting behavior
I just noticed that the various vertico-directory commands and MCT's
DEL binding non actually behave like my old up-directory which deletes
the last path component whether or not there is a slash at the end.
I'm going to try to get used to using / DEL in that case, but if I
can't get used to it I'll resurrect up-directory.