Make define-case-changer like transpose-chars
Previously the commands defined by define-case-changer would work on
the previous word if point was at the end of a word but the previous
command was not one of my case-change commands. Now they have a new
simpler definition: they work on the previous word if at the end of
line, just like transpose-chars does.