Fix other-window annoyance
I keep trying to switch to another window and type a word that begins
with 'o' and getting very confused because I wind up not typing an 'o'
and typing what comes after 'o' in a different window altogether!

The M-o keybinding is short enough to not need a repeat map. Also, I
almost never have more than 3 windows in a frame, so bashing M-o is
OK, I don't need akey to go backwards.