Give jinx a serious try
Even if the squiggles prove to be to aggressive for me, I could still
try toggling the mode on and off. It is pretty smooth in a quick test.

Maybe what I really need is a personal word list. I've never tried
that before (I always just accept for a single session), and
spell-checking always has tons of false positives for technical
terms and proper names.

ispell-region is bound in embark-region-map, but I think I don't use
it, so I really will be switching over completely to jinx for the
duration of this experiment.