Add key binding for consult-line-multi
I just tried this thanks to a comment of @minad's on the Consult
wishlist issue and it looks really nice to switch buffers based on
contents.  I mostly use consult-imenu-multi for that, but there are
obviously some cases it would not cover that consult-line-multi does.

Also, as @minad pointed out in that issue, consult-line-multi only
allocates lines that match the search portion, whereas consult-line
gathers all lines regardless of the input.

This commit also removes the key binding for consult-multi-occur which
seems to no longer exist (I never used it anyway, I think). Maybe it
turned into this consult-line-multi I just learned about?