Add binding for new duplicate-dwim command
I usually duplicate a region with embark-insert and a line with
copy-from-above-command, but this new command can do both. Let's see
if I find it more convenient or not. With no active region, it
duplicates the current line, not the one above, so it'll take some
getting used to. Also, with an active region, it keeps it active,
which will also take some getting used to. Maybe this is just not
for me. 😀