Use cycle-spacing instead of delete-horizontal-space
I was used to having just-one-space bound to M-SPC and
delete-horizontal-space on M-\, but I cannot use M-SPC on WSL, so
there I was using only delete-horizontal-space and M-\ SPC if I needed
a space. But I just noticed there is a cycle-spacing-actions variable
that lets me configure cycle-spacing to first behave as
delete-all-space and then as just-one-space (the default is the other
way around). So I can bind cycle-spacing to M-\ and now have M-\ M-\
as an alternative to M-\ SPC.