Convert this-buffer embark target to a command
I had a target finder for the current buffer that I gave lowest
priority. That way it didn't get much in the way of normal embark-act
use, but C-- C-. acted as a prefix for actions on the current buffer.
But sometimes it would be nice if embark-act had no targets at all,
for example, to use embark eldoc functions. So I've converted that
target finder to its own command (that let-binds the target finder for
the current buffer and calls embark-act).