Turn off org-modern fold indicators
A recent update to org-modern changed the heading bullets to triangles
to indicate folding. I think I might like that in the long run, but
for now I'll turn it off because the characters chosen by default
include some tofus in my fonts and also differ greatly in size between
the folded and unfolded characters for the same heading. I bet I can
find some suitable characters for the folds, but I don't want to do it
right now.